requirements.txt
arrow==1.3.0
asgiref==3.8.1
async-timeout==4.0.3
blessed==1.20.0
Django==3.2
django-picklefield==3.2
django-q==1.3.9
django-redis==5.4.0
djangorestframework==3.12.4
pillow==10.4.0
python-dateutil==2.9.0.post0
pytz==2024.2
redis==3.5.3
six==1.16.0
sqlparse==0.5.1
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
wcwidth==0.2.13
